Add files via upload
A city wants to set up a district heating system and the authority provides us with demand for annual (January/01/2018 to December/31/2018) heating. The main source of power supply is the PV module and the electric main grid. The main components installed for the heating system are CHP, an electric-operated ambient air heat pump, natural gas boiler, and thermal storage.
Open Energy Modeling Framework(oemof) is used to model an energy system. As per the data of CAPEX, we performed time schedule to calculate the optimum use of components at different times based on the pricing of electricity, natural gas, and co2 emission. The use of green hydrogen also uses in comparison with natural gas.